Improve the data fetching logic
