+++ src/components/02.Organisms/Project/project-item.jsx
@@ -1,24 +1,17 @@
 import Image from 'next/image';
-import apiService from '@/app/utils/service';
 import Title from '@/components/01.Atoms/Title/title';
 import SocialButton from '@/components/01.Atoms/Socials/social-button';
 import Animate from '@/components/01.Atoms/MountTransition/fader';
 import TechStack from '@/components/02.Organisms/Stack/tech-stack';
 import { faGithub } from '@fortawesome/free-brands-svg-icons';
+import getMedia from '@/app/lib/getMedia';
 
 import styles from '../../03.Molecules/ProjectsList/projects-list.module.scss';
 
-async function getMedia(image) {
-  const data = await apiService.getThumbnail(image);
-
-  return {
-    sizes: data.sizes,
-    url: data.source_url,
-  }
-}
-
 const ProjectItem = async ({ className, title, description, year, stack, link, image }) => {
-  const thumbnail = await getMedia(image);
+  const media = await getMedia(image);
+
+  const [thumbnail] = await Promise.all([media]);
 
   return (
     <div className={`${styles.projectItem} ${className}`}>
@@ -49,7 +42,7 @@ const ProjectItem = async ({ className, title, description, year, stack, link, i
         </div>
       </div>
 
-      {!!thumbnail && (
+      {thumbnail && (
         <figure className={styles.projectBackground}>
           <Image fill={true} className={styles.background} src={thumbnail.url} alt="" />
         </figure>
