+++ src/components/03.Molecules/ProjectsList/projects-list.jsx
@@ -1,27 +1,13 @@
-import apiService from '@/app/utils/service';
+import getProjectsList from '@/app/lib/getProjects';
 import Animate from '@/components/01.Atoms/MountTransition/fader';
 import ProjectItem from '@/components/02.Organisms/Project/project-item';
-import he from 'he';
 
 import styles from './projects-list.module.scss'
 
-async function getData() {
-  const data = await apiService.getProjects();
-  const formattedPosts = data.map(post => ({
-    id: post.id,
-    title: he.decode(post.title.rendered),
-    content: post.the_content,
-    tags: post.tags,
-    repository: post.repository,
-    endYear: post.end_year,
-    thumbnailId: post.featured_media,
-  }));
-
-  return formattedPosts;
-}
-
 const ProjectsList = async () => {
-  const posts = await getData();
+  const projectsList = await getProjectsList();
+
+  const [posts] = await Promise.all([projectsList]);
 
   return (
     <div className={styles.projectsList}>
